+++ Source/Scene/findTileMetadata.js
@@ -1,6 +1,8 @@
+import defaultValue from "../Core/defaultValue.js";
 import defined from "../Core/defined.js";
 import hasExtension from "./hasExtension.js";
 import TileMetadata from "./TileMetadata.js";
+import oneTimeWarning from "../Core/oneTimeWarning.js";
 
 /**
  * Check if a tile has metadata, either defined in its metadata field (3D Tiles 1.1)
@@ -17,7 +19,7 @@ import TileMetadata from "./TileMetadata.js";
  * @private
  * @experimental This feature is using part of the 3D Tiles spec that is not final and is subject to change without Cesium's standard deprecation policy.
  */
-export default function findTileMetadata(tileset, tileHeader) {
+function findTileMetadata(tileset, tileHeader) {
   const metadataJson = hasExtension(tileHeader, "3DTILES_metadata")
     ? tileHeader.extensions["3DTILES_metadata"]
     : tileHeader.metadata;
@@ -26,7 +28,18 @@ export default function findTileMetadata(tileset, tileHeader) {
     return undefined;
   }
 
-  const classes = tileset.schema.classes;
+  if (!defined(tileset.schema)) {
+    findTileMetadata._oneTimeWarning(
+      "findTileMetadata-missing-root-schema",
+      "Could not find a metadata schema for tile metadata. For tilesets that contain external tilesets, make sure the schema is added to the root tileset.json."
+    );
+    return undefined;
+  }
+
+  const classes = defaultValue(
+    tileset.schema.classes,
+    defaultValue.EMPTY_OBJECT
+  );
   if (defined(metadataJson.class)) {
     const tileClass = classes[metadataJson.class];
     return new TileMetadata({
@@ -37,3 +50,7 @@ export default function findTileMetadata(tileset, tileHeader) {
 
   return undefined;
 }
+
+// Exposed for testing
+findTileMetadata._oneTimeWarning = oneTimeWarning;
+export default findTileMetadata;