+++ Specs/DataSources/ModelVisualizerSpec.js
@@ -176,7 +176,7 @@ describe(
       expect(primitive.clippingPlanes._planes[0].distance).toEqual(
         clippingPlanes._planes[0].distance
       );
-      expect(primitive.imageBasedLightingFactor).toEqual(
+      expect(primitive.imageBasedLighting.imageBasedLightingFactor).toEqual(
         new Cartesian2(0.5, 0.5)
       );
       expect(primitive.lightColor).toEqual(new Color(1.0, 1.0, 0.0, 1.0));