+++ Specs/Scene/ResourceCacheSpec.js
@@ -247,20 +247,14 @@ describe(
       ],
     };
 
-    // Index buffers will load as typed arrays only in WebGL1.
-    // In order to load them as buffers, the scene must have WebGL2 enabled.
     let scene;
-    let sceneWithWebgl2;
 
     beforeAll(function () {
       scene = createScene();
-      sceneWithWebgl2 = createScene();
-      sceneWithWebgl2.context._webgl2 = true;
     });
 
     afterAll(function () {
       scene.destroyForSpecs();
-      sceneWithWebgl2.destroyForSpecs();
     });
 
     afterEach(function () {
@@ -857,6 +851,7 @@ describe(
         gltfResource: gltfResource,
         baseResource: gltfResource,
         bufferViewId: 0,
+        loadBuffer: true,
       });
       const vertexBufferLoader = ResourceCache.loadVertexBuffer({
         gltf: gltfUncompressed,
@@ -864,6 +859,7 @@ describe(
         baseResource: gltfResource,
         bufferViewId: 0,
         accessorId: 0,
+        loadBuffer: true,
       });
 
       const cacheEntry = ResourceCache.cacheEntries[expectedCacheKey];
@@ -877,6 +873,7 @@ describe(
           gltfResource: gltfResource,
           baseResource: gltfResource,
           bufferViewId: 0,
+          loadBuffer: true,
         })
       ).toBe(vertexBufferLoader);
 
@@ -904,6 +901,7 @@ describe(
         baseResource: gltfResource,
         draco: dracoExtension,
         attributeSemantic: "POSITION",
+        loadBuffer: true,
       });
       const vertexBufferLoader = ResourceCache.loadVertexBuffer({
         gltf: gltfDraco,
@@ -912,6 +910,7 @@ describe(
         draco: dracoExtension,
         attributeSemantic: "POSITION",
         accessorId: 0,
+        loadBuffer: true,
       });
 
       const cacheEntry = ResourceCache.cacheEntries[expectedCacheKey];
@@ -927,6 +926,7 @@ describe(
           draco: dracoExtension,
           attributeSemantic: "POSITION",
           accessorId: 0,
+          loadBuffer: true,
         })
       ).toBe(vertexBufferLoader);
 
@@ -949,7 +949,7 @@ describe(
         gltfResource: gltfResource,
         baseResource: gltfResource,
         bufferViewId: 0,
-        loadAsTypedArray: true,
+        loadTypedArray: true,
       });
       const vertexBufferLoader = ResourceCache.loadVertexBuffer({
         gltf: gltfUncompressed,
@@ -957,7 +957,7 @@ describe(
         baseResource: gltfResource,
         bufferViewId: 0,
         accessorId: 0,
-        loadAsTypedArray: true,
+        loadTypedArray: true,
       });
 
       expect(vertexBufferLoader.cacheKey).toBe(expectedCacheKey);
@@ -970,6 +970,39 @@ describe(
       });
     });
 
+    it("loads vertex buffer as buffer and typed array for 2D", function () {
+      spyOn(Resource.prototype, "fetchArrayBuffer").and.returnValue(
+        Promise.resolve(bufferArrayBuffer)
+      );
+
+      const expectedCacheKey = ResourceCacheKey.getVertexBufferCacheKey({
+        gltf: gltfUncompressed,
+        gltfResource: gltfResource,
+        baseResource: gltfResource,
+        bufferViewId: 0,
+        loadBuffer: true,
+        loadTypedArray: true,
+      });
+      const vertexBufferLoader = ResourceCache.loadVertexBuffer({
+        gltf: gltfUncompressed,
+        gltfResource: gltfResource,
+        baseResource: gltfResource,
+        bufferViewId: 0,
+        accessorId: 0,
+        loadBuffer: true,
+        loadTypedArray: true,
+      });
+
+      expect(vertexBufferLoader.cacheKey).toBe(expectedCacheKey);
+
+      return waitForLoaderProcess(vertexBufferLoader, scene).then(function (
+        vertexBufferLoader
+      ) {
+        expect(vertexBufferLoader.typedArray).toBeDefined();
+        expect(vertexBufferLoader.buffer).toBeDefined();
+      });
+    });
+
     it("loadVertexBuffer throws if gltf is undefined", function () {
       expect(function () {
         ResourceCache.loadVertexBuffer({
@@ -977,6 +1010,7 @@ describe(
           gltfResource: gltfResource,
           baseResource: gltfResource,
           bufferViewId: 0,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -988,6 +1022,7 @@ describe(
           gltfResource: undefined,
           baseResource: gltfResource,
           bufferViewId: 0,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -999,6 +1034,7 @@ describe(
           gltfResource: gltfResource,
           baseResource: undefined,
           bufferViewId: 0,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -1013,6 +1049,7 @@ describe(
           draco: dracoExtension,
           attributeSemantic: "POSITION",
           accessorId: 0,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -1023,6 +1060,7 @@ describe(
           gltf: gltfDraco,
           gltfResource: gltfResource,
           baseResource: gltfResource,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -1036,6 +1074,7 @@ describe(
           draco: dracoExtension,
           attributeSemantic: undefined,
           accessorId: 0,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -1049,6 +1088,20 @@ describe(
           draco: dracoExtension,
           attributeSemantic: "POSITION",
           accessorId: undefined,
+          loadBuffer: true,
+        });
+      }).toThrowDeveloperError();
+    });
+
+    it("loadVertexBuffer throws if both loadBuffer and loadTypedArray are false", function () {
+      expect(function () {
+        ResourceCache.loadVertexBuffer({
+          gltf: gltfUncompressed,
+          gltfResource: gltfResource,
+          baseResource: gltfResource,
+          bufferViewId: 0,
+          loadBuffer: false,
+          loadTypedArray: false,
         });
       }).toThrowDeveloperError();
     });
@@ -1063,12 +1116,14 @@ describe(
         accessorId: 2,
         gltfResource: gltfResource,
         baseResource: gltfResource,
+        loadBuffer: true,
       });
       const indexBufferLoader = ResourceCache.loadIndexBuffer({
         gltf: gltfUncompressed,
         accessorId: 2,
         gltfResource: gltfResource,
         baseResource: gltfResource,
+        loadBuffer: true,
       });
 
       const cacheEntry = ResourceCache.cacheEntries[expectedCacheKey];
@@ -1082,6 +1137,7 @@ describe(
           accessorId: 2,
           gltfResource: gltfResource,
           baseResource: gltfResource,
+          loadBuffer: true,
         })
       ).toBe(indexBufferLoader);
 
@@ -1091,6 +1147,7 @@ describe(
         indexBufferLoader
       ) {
         expect(indexBufferLoader.buffer).toBeDefined();
+        expect(indexBufferLoader.typedArray).toBeUndefined();
       });
     });
 
@@ -1109,6 +1166,7 @@ describe(
         gltfResource: gltfResource,
         baseResource: gltfResource,
         draco: dracoExtension,
+        loadBuffer: true,
       });
       const indexBufferLoader = ResourceCache.loadIndexBuffer({
         gltf: gltfDraco,
@@ -1116,6 +1174,7 @@ describe(
         gltfResource: gltfResource,
         baseResource: gltfResource,
         draco: dracoExtension,
+        loadBuffer: true,
       });
 
       const cacheEntry = ResourceCache.cacheEntries[expectedCacheKey];
@@ -1130,19 +1189,20 @@ describe(
           gltfResource: gltfResource,
           baseResource: gltfResource,
           draco: dracoExtension,
+          loadBuffer: true,
         })
       ).toBe(indexBufferLoader);
 
       expect(cacheEntry.referenceCount).toBe(2);
 
-      return waitForLoaderProcess(indexBufferLoader, sceneWithWebgl2).then(
-        function (indexBufferLoader) {
-          expect(indexBufferLoader.buffer).toBeDefined();
-        }
-      );
+      return waitForLoaderProcess(indexBufferLoader, scene).then(function (
+        indexBufferLoader
+      ) {
+        expect(indexBufferLoader.buffer).toBeDefined();
+      });
     });
 
-    it("loads index buffer as typed array using option", function () {
+    it("loads index buffer as typed array", function () {
       spyOn(Resource.prototype, "fetchArrayBuffer").and.returnValue(
         Promise.resolve(bufferArrayBuffer)
       );
@@ -1152,14 +1212,14 @@ describe(
         accessorId: 2,
         gltfResource: gltfResource,
         baseResource: gltfResource,
-        loadAsTypedArray: true,
+        loadTypedArray: true,
       });
       const indexBufferLoader = ResourceCache.loadIndexBuffer({
         gltf: gltfUncompressed,
         accessorId: 2,
         gltfResource: gltfResource,
         baseResource: gltfResource,
-        loadAsTypedArray: true,
+        loadTypedArray: true,
       });
 
       expect(indexBufferLoader.cacheKey).toBe(expectedCacheKey);
@@ -1170,81 +1230,6 @@ describe(
         expect(indexBufferLoader.typedArray).toBeDefined();
         expect(indexBufferLoader.buffer).toBeUndefined();
       });
-    });
-
-    it("loads index buffer as typed array for wireframes in WebGL1", function () {
-      spyOn(Resource.prototype, "fetchArrayBuffer").and.returnValue(
-        Promise.resolve(bufferArrayBuffer)
-      );
-
-      const expectedCacheKey = ResourceCacheKey.getIndexBufferCacheKey({
-        gltf: gltfUncompressed,
-        accessorId: 2,
-        gltfResource: gltfResource,
-        baseResource: gltfResource,
-        loadAsTypedArray: false,
-      });
-      const indexBufferLoader = ResourceCache.loadIndexBuffer({
-        gltf: gltfUncompressed,
-        accessorId: 2,
-        gltfResource: gltfResource,
-        baseResource: gltfResource,
-        loadAsTypedArray: false,
-        loadForWireframe: true,
-      });
-
-      expect(indexBufferLoader.cacheKey).toBe(expectedCacheKey);
-
-      return waitForLoaderProcess(indexBufferLoader, scene).then(function (
-        indexBufferLoader
-      ) {
-        expect(indexBufferLoader.typedArray).toBeDefined();
-        expect(indexBufferLoader.buffer).toBeUndefined();
-      });
-    });
-
-    it("loads index buffer as buffer for wireframes in WebGL2", function () {
-      spyOn(Resource.prototype, "fetchArrayBuffer").and.returnValue(
-        Promise.resolve(bufferArrayBuffer)
-      );
-
-      const expectedCacheKey = ResourceCacheKey.getIndexBufferCacheKey({
-        gltf: gltfUncompressed,
-        accessorId: 2,
-        gltfResource: gltfResource,
-        baseResource: gltfResource,
-        loadForWireframe: true,
-      });
-      const indexBufferLoader = ResourceCache.loadIndexBuffer({
-        gltf: gltfUncompressed,
-        accessorId: 2,
-        gltfResource: gltfResource,
-        baseResource: gltfResource,
-        loadForWireframe: true,
-      });
-
-      const cacheEntry = ResourceCache.cacheEntries[expectedCacheKey];
-      expect(indexBufferLoader.cacheKey).toBe(expectedCacheKey);
-      expect(cacheEntry.referenceCount).toBe(1);
-
-      // The existing resource is returned if the computed cache key is the same
-      expect(
-        ResourceCache.loadIndexBuffer({
-          gltf: gltfUncompressed,
-          accessorId: 2,
-          gltfResource: gltfResource,
-          baseResource: gltfResource,
-        })
-      ).toBe(indexBufferLoader);
-
-      expect(cacheEntry.referenceCount).toBe(2);
-
-      return waitForLoaderProcess(indexBufferLoader, sceneWithWebgl2).then(
-        function (indexBufferLoader) {
-          expect(indexBufferLoader.typedArray).toBeUndefined();
-          expect(indexBufferLoader.buffer).toBeDefined();
-        }
-      );
     });
 
     it("loadIndexBuffer throws if gltf is undefined", function () {
@@ -1254,6 +1239,7 @@ describe(
           accessorId: 2,
           gltfResource: gltfResource,
           baseResource: gltfResource,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -1265,6 +1251,7 @@ describe(
           accessorId: undefined,
           gltfResource: gltfResource,
           baseResource: gltfResource,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -1276,6 +1263,7 @@ describe(
           accessorId: 2,
           gltfResource: undefined,
           baseResource: gltfResource,
+          loadBuffer: true,
         });
       }).toThrowDeveloperError();
     });
@@ -1287,6 +1275,20 @@ describe(
           accessorId: 2,
           gltfResource: gltfResource,
           baseResource: undefined,
+          loadBuffer: true,
+        });
+      }).toThrowDeveloperError();
+    });
+
+    it("loadIndexBuffer throws if both loadBuffer and loadTypedArray are false", function () {
+      expect(function () {
+        ResourceCache.loadIndexBuffer({
+          gltf: gltfUncompressed,
+          accessorId: 2,
+          gltfResource: gltfResource,
+          baseResource: gltfResource,
+          loadBuffer: false,
+          loadTypedArray: false,
         });
       }).toThrowDeveloperError();
     });